+++ frontend/src/components/CardImg/styles.css
@@ -13,12 +13,13 @@
 
 .imagem{
     width: 100%;
-    height: 100%;
+    height: auto;
     object-fit: contain;
     display: flex;
     justify-content: center;
     align-items: center;
     border-radius: 10px;
+   
 }
 
 .box-imagem{
@@ -32,4 +33,9 @@
     justify-content: center;
     align-items: center;
     margin-top: -20px;
+}
+
+.box-imagem img {
+    box-shadow: 8px 8px 1px 0px #00000040;
+
 }
